tests: move golog test fixture into a separate library
This allows us to re-use it for additional tests.
